pyproject.toml
[project]
name = "rag-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mailbox>=0.4",
    "textual>=1.0.0",
    "textual-dev>=1.7.0",
    "faiss-cpu>=1.9.0.post1",
    "beautifulsoup4>=4.12.3",
    "kivy>=2.3.1",
    "kivymd>=1.2.0",
    "sentence-transformers>=3.3.1",
    "chroma>=0.2.0",
    "numpy>=2.2.0",
    "scikit-learn>=1.6.0",
    "python-dotenv>=1.0.1",
    "openai>=1.59.7",
    "langchain>=0.3.14",
    "o365>=2.0.38",
    "langchain-community>=0.3.14",
    "langchain-openai>=0.3.0",
]
